pyproject.toml
[project]
name = "matrixrender"
version = "0.1.0"
description = "Turn a numpy matrix into a rendered screen via PyGame"
authors = [
    {name = "Sandu Bogdan",email = "bogdanelsandu2011@gmail.com"}
]
license = {text = "MIT License"}
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy (>=2.4.1,<3.0.0)",
    "pygame (>=2.6.1,<3.0.0)",
    "pillow (>=12.1.0,<13.0.0)"
]

[dependency-groups]
dev = [
    "ruff (>=0.14.14,<0.15.0)",
    "mypy (>=1.19.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
